+++ core/templates/dev/head/pages/exploration_editor/exploration_editor.html
@@ -92,7 +92,7 @@
       </div>
 
       <div ng-show="getActiveTabName() === 'feedback'">
-        {% include 'pages/exploration_editor/feedback_tab/feedback_tab.html' %}
+        <feedback-tab></feedback-tab>
       </div>
     </div>
   </div>
