--- core/templates/dev/head/pages/exploration_editor/feedback_tab/feedback_tab.html
@@ -1,156 +1,0 @@
-<md-card class="oppia-editor-card oppia-feedback-card" ng-controller="FeedbackTab">
-  <div ng-if="!activeThread">
-    <div ng-if="threadData.suggestionThreads.length > 0">
-      <h2>Suggestions from Learners</h2>
-      <thread-table threads="threadData.suggestionThreads"
-                    on-click-row="setActiveThread">
-      </thread-table>
-      <br>
-    </div>
-
-    <div class="pull-right">
-      <button ng-if="userIsLoggedIn" class="btn btn-success pull-right" ng-click="showCreateThreadModal()">
-        Start new thread
-      </button>
-    </div>
-
-    <h2>Exploration Feedback</h2>
-
-    <div style="clear: both;"></div>
-
-    <div ng-if="threadData.feedbackThreads.length === 0">
-      <em>No feedback has been given for this exploration yet.</em>
-    </div>
-
-    <thread-table threads="threadData.feedbackThreads"
-                  on-click-row="setActiveThread">
-    </thread-table>
-
-    <em ng-if="!userIsLoggedIn">To create a new feedback thread, please log in.</em>
-  </div>
-
-  <div ng-if="activeThread">
-    <div class="row">
-      <div class="col-lg-9 col-md-9 col-sm-9">
-        <span class="oppia-back-arrow">
-          <button type="button" class="btn btn-default btn-xs protractor-test-oppia-feedback-back-button" ng-click="onBackButtonClicked()">
-            <i class="material-icons oppia-vcenter" title="Return to list of feedback threads">&#xE5CB;</i>
-          </button>
-        </span>
-        <span style="font-size: larger" ng-show="!activeThread.isSuggestionThread()">Feedback Thread: <[activeThread.subject]></span>
-        <span style="font-size: larger" ng-show="activeThread.isSuggestionThread()">Suggestion Thread: <[activeThread.subject]></span>
-      </div>
-
-      <div style="float: right;" class="col-lg-5 col-md-5 col-sm-5">
-        <div class="pull-right">
-          <span ng-if="activeThread.stateName" class="label label-info">
-            state: <[activeThread.stateName]>
-          </span>
-          <span ng-class="getLabelClass(activeThread.status)">
-            <[getHumanReadableStatus(activeThread.status)]>
-          </span>
-        </div>
-      </div>
-    </div>
-
-    <div class="row" ng-if="activeThread.messages">
-      <div class="col-lg-12 col-md-12 col-sm-12 activeThreadView">
-        <table class="table">
-          <tr ng-repeat="m in activeThread.messages|orderBy:'message_id'">
-            <td>
-              <div class="row">
-                <div class="col-lg-5 col-md-5 col-sm-5">
-                  <span ng-if="$index !== 0">#<[m.message_id]></span>
-                  <span ng-if="m.author_username">by <strong><[m.author_username]></strong></span>
-                  <span ng-if="!m.author_username">(anonymously submitted)</span>
-                </div>
-
-                <div class="col-lg-4 col-md-4 col-sm-4">
-                  <span ng-if="m.message_id !== 0">
-                    <span ng-if="m.updated_status">
-                      <em>Status changed to '<[getHumanReadableStatus(m.updated_status)]>'</em>
-                    </span>
-                    <span ng-if="m.updated_subject">
-                      <em>Subject changed to '<[m.updated_subject]>'</em>
-                    </span>
-                  </span>
-                </div>
-
-                <div class="col-lg-3 col-md-3 col-sm-3">
-                  <span><[getLocaleAbbreviatedDatetimeString(m.created_on)]></span>
-                </div>
-              </div>
-
-              <div class="row">
-                <div class="col-lg-12 col-md-12 col-sm-12">
-                  <div style="white-space: pre-wrap;" class="protractor-test-exploration-feedback"><[m.text]></div>
-                  <div ng-if="activeThread.isSuggestionThread() && $index == 0">
-                    <button class="btn btn-<[getSuggestionButtonType()]> protractor-test-view-suggestion-btn" style="margin-top: 6px; margin-bottom: 6px" ng-click="showSuggestionModal()">
-                      View Suggestion
-                    </button>
-                  </div>
-                </div>
-              </div>
-            </td>
-          </tr>
-        </table>
-      </div>
-    </div>
-
-
-    <hr>
-
-    <div class="row" ng-if="userIsLoggedIn">
-      <div class="col-lg-12 col-md-12 col-sm-12">
-        <div>
-          <label for="tmpMessageText">
-            Add new message
-            <span ng-if="EditabilityService.isEditable() && !activeThread.isSuggestionThread()">(and/or change status)</span>
-          </label>
-          <textarea class="form-control protractor-test-feedback-response-textarea" ng-model="tmpMessage.text"
-                    id="tmpMessageText" rows="6"
-                    ng-disabled="messageSendingInProgress">
-          </textarea>
-        </div>
-
-        <div>
-          <span ng-show="EditabilityService.isEditable() && !activeThread.isSuggestionThread()">
-            Change status (optional):
-            <select ng-model="tmpMessage.status"
-                    ng-options="choice.id as choice.text for choice in STATUS_CHOICES"
-                    ng-disabled="messageSendingInProgress">
-            </select>
-            <div style="color:#ff0000; margin-top: 7px;" ng-if="!tmpMessage.text && (tmpMessage.status === 'ignored' || tmpMessage.status === 'not_actionable')">
-              <i class="material-icons">&#xE88F;</i>
-              <span>Please specify a reason for setting the status to <[getHumanReadableStatus(tmpMessage.status)]>.</span>
-            </div>
-          </span>
-        </div>
-
-        <div>
-          <button class="btn btn-success protractor-test-oppia-feedback-response-send-btn" style="margin-top: 10px;"
-                  ng-click="addNewMessage(activeThread.threadId, tmpMessage.text, tmpMessage.status)"
-                  ng-disabled="messageSendingInProgress || (!tmpMessage.text && activeThread.status === tmpMessage.status) || (!tmpMessage.text && (tmpMessage.status === 'ignored' || tmpMessage.status === 'not_actionable'))">
-            <span ng-if="messageSendingInProgress">Sending...</span>
-            <span ng-if="!messageSendingInProgress">Send</span>
-          </button>
-        </div>
-      </div>
-    </div>
-    <em ng-if="!userIsLoggedIn">To respond to a feedback thread, please log in.</em>
-  </div>
-</md-card>
-
-
-<style>
-  /* Overwrite the default "success" color to a darker shade of green in order
-  to prevent this overshadowing the active action buttons (like "Create New
-  Thread"). */
-  .oppia-feedback-card span.label.label-success {
-    background-color: #038603;
-  }
-  .activeThreadView {
-    margin-top: 10px;
-    word-wrap: break-word;
-  }
-</style>
